# swagger.yaml
openapi: 3.0.3
info:
  title: Notes API - OpenAPI 3.0
  version: 1.0.1
  description: |-
    This is a sample services based on the OpenAPI 3.0 specification. You can make short notes(404 symbols)
tags:
  - name: user
    description: Operations about user
  - name: notes
    description: Operations about note
  - name: tag
    description: Operations about tag
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can be done by not logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserInfo'


  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: true
          schema:
            type: string
          example: theUser
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      description: ''
      operationId: logoutUser
      parameters: []
      responses:
        default:
          description: successful operation
  /user/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
          example: theUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          description: Unknown error
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        default:
          description: successful operation
      security:
        - note_auth:
          - admin
          - user
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
      security:
        - note_auth:
          - admin
          - user
  /user/stat/{username}:
    get:
      tags:
        - user
      summary: Get user stats by user id
      description: ''
      operationId: getStatsById
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use 10 for testing. '
          required: true
          schema:
            type: string
          example: theUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
            application/xml:
              schema:
                $ref: '#/components/schemas/Stats'
        '400':
          description: Invalid user id supplied
        '404':
          description: Stats for user not found
        default:
          description: Unknown error
      security:
        - note_auth:
          - admin
          - user
  /notes:
    post:
      tags:
        - notes
      summary: Create note
      description: This can only be done by the logged in user.
      operationId: createNote
      requestBody:
        description: Created note object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notes'
          application/xml:
            schema:
              $ref: '#/components/schemas/Notes'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Notes'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
        default:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
      security:
        - note_auth:
          - admin
          - user
  /notes/{id}:
    get:
      tags:
        - notes
      summary: Get note by id
      description: ''
      operationId: getNoteById
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use 1 for testing. '
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
        '400':
          description: Invalid id supplied
        '404':
          description: Note not found
        default:
          description: Unknown error
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - notes
      summary: Update note
      description: This can only be done by the logged in user.
      operationId: updateNotes
      parameters:
        - name: id
          in: path
          description: id that need to be deleted
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent note
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
          application/xml:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
      responses:
        default:
          description: successful operation
      security:
        - note_auth:
          - admin
          - user

    delete:
      tags:
        - notes
      summary: Delete note
      description: This can only be done by the logged in user.
      operationId: deleteNotes
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '400':
          description: Invalid id supplied
        '404':
          description: Note not found
      security:
        - note_auth:
          - admin
          - user
  /tag:
    post:
      tags:
        - tag
      summary: Create tag
      description: This can only be done by the logged in user.
      operationId: createTag
      requestBody:
        description: Created tag object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        default:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
      security:
        - note_auth:
          - admin
          - user
  /tag/{id}:
    get:
      tags:
        - tag
      summary: Get tag by id
      description: ''
      operationId: getTagById
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use 1 for testing. '
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid id supplied
        '404':
          description: Note not found
        default:
          description: Unknown error
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - tag
      summary: Update tag
      description: This can only be done by the admin
      operationId: updateTag
      parameters:
        - name: id
          in: path
          description: id that need to be updated
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent note
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        default:
          description: successful operation
      security:
        - note_auth:
          - admin
    delete:
      tags:
        - tag
      summary: Delete tag
      description: This can only be done by the admin
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '400':
          description: Invalid id supplied
        '404':
          description: Note not found
      security:
        - note_auth:
          - admin
  /notes/allowed:
    post:
      tags:
        - notes
      summary: allow a editing
      description: This can only be done by the logged in user.
      operationId: allowEditing
      requestBody:
        description: Created note object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/allowedNote'
          application/xml:
            schema:
              $ref: '#/components/schemas/allowedNote'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/allowedNote'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
        default:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
            application/xml:
              schema:
                $ref: '#/components/schemas/Notes'
      security:
        - note_auth:
          - admin
          - user

components:
  schemas:
    UserInfo:
      type: object
      required:
      - id
      - email
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: theUser
        email:
          type: string
          example: ivan@email.com
        firstName:
          type: string
          example: Ivan
        lastName:
          type: string
          example: James
        userStatus:
          type: string
          enum:
            - online
            - offline
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Notes'
        stats:
          $ref: '#/components/schemas/Stats'
    UserCreate:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          type: string
          example: theUser
        email:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
          example: theUser
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
      xml:
        name: user
    Notes:
      type: object
      required:
        - id
        - ownerId
        - isPublic
      properties:
        id:
          type: integer
          format: int64
          example: 10
        ownerId:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: 'My title'
        isPublic:
          type: boolean
          example: true
        text:
          type: string
          example: 'Today i was in Lviv. It was great!'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/UserInfo'
        dateOfEditing:
          type: array
          items:
            type: string
            format: date
            example: '2022-10-02'
      xml:
        name: notes
    NotesUpdate:
      type: object
      properties:
        title:
          type: string
          example: 'My title'
        isPublic:
          type: boolean
          example: true
        text:
          type: string
          example: 'Today i was in Lviv. It was great!'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagUpdate'
      xml:
        name: notesUp
    Tag:
      type: object
      required:
        - id
        - text
      properties:
        id:
          type: integer
          format: int64
          example: 3
        text:
          type: string
          example: '#lviv'
      xml:
        name: tag
    TagUpdate:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          example: '#lviv'
      xml:
        name: tagUpdate
    Stats:
      type: object
      required:
        - id
        - userId
      properties:
        id:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 10
        numOfNotes:
          type: integer
          format: int64
          example: 9
        numOfEditingNotes:
          type: integer
          format: int64
          example: 9
        dateOfCreating:
          type: string
          format: date
          example: '2022-10-02'
      xml:
        name: stats
    allowedNote:
      type: object
      properties:
        noteId:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 10
      xml:
        name: allowedNotes
  requestBodies:
    UserArray:
      description: List of user object
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/UserCreate'
  securitySchemes:
    note_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: url
          scopes:
            admin : can do all actions
            user : can do only allowed action
    api_key:
      type: apiKey
      name: api_key
      in: header